requirements.txt
# Web Frameworks
streamlit>=1.44.0
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.0.0
gunicorn>=21.2.0

# LangChain Ecosystem
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-google-genai>=2.0.0
langchain-chroma>=0.1.0
langchain-groq>=0.1.0

# Database & Processing
chromadb>=0.5.0
pymupdf>=1.24.0      # Required for fitz in Extract.py
python-dotenv>=1.0.0

# Optional — Hybrid / Reranking support
rank-bm25>=0.2.2
sentence-transformers>=3.0.0
